rag_pipeline/requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
pypdf2>=3.0.0
python-docx>=0.8.11
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0